requirements.txt
alembic==1.0.6
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kazoo==2.5.0
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
pykafka==2.8.0
pyral==1.4.2
pytest==4.1.1
python-dateutil==2.7.5
python-editor==1.0.3
PyYAML==3.13
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.16
tabulate==0.8.3
urllib3==1.24.1
Werkzeug==0.14.1
